Fix the rmmod failed issue in my modules.
Previous version has an issue of `rmmod`.

If you `rmmod` then linux saids that you are trying to remove
the built-in module.

I didn't know why at first. But I checked the `modules.buitin` file
which is located in `/lib/modules/$(uname -r)/` and I saw the `pblk`
driver in that file.

That is not file that is really built-in file which is located in
linux kernel's `/driver/lightnvm`.

In other words, my module name and original pblk module name are crashed
by each other.

As a result, I change the module name `pblk` to `mypblk`.